fix: evita que las rutas  client-only rompan el build de  Next
