+++ next.config.mjs
@@ -1,8 +1,6 @@
 /**
  * @type {import("next").NextConfig}
  */
-const config = {
-  swcMinify: false,
-};
+const config = {};
 
 export default config;
